LPS-94205 Make PortletToolbar unit test
