--- portal-impl/test/integration/com/liferay/portal/kernel/portlet/toolbar/bundle/portlettoolbar/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Description: Test Bundle for PortletToolbarTest
-Bundle-Version: 1.0.0
-Private-Package: $(bundle.package).*
--dsannotations: *
